Separates reporting calls into its own service, fixes add to reporting ui
